Extract StateStoreCommitterRunsBuilder
